requirements.txt
Django==1.8.2
django-tinymce==2.2.0
pillow==3.1.1
